requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob>=12.0.0
pandas>=2.0.0
openpyxl>=3.0.0
xlsxwriter>=3.0.0